clippy fixes

+++ benches/reader_bench.rs
@@ -7,9 +7,9 @@ fn bench_reader(c: &mut Criterion) {
     c.bench_function("reader_read_bits_1", |b| {
         b.iter(|| {
             let mut reader = Reader::new(black_box(&data), 0, data.len());
-            let mut sum = 0;
+            let mut sum = 0u64;
             for _ in 0..1000 {
-                sum += reader.read_bits(1).unwrap_or(0);
+                sum += u64::from(reader.read_bits(1).unwrap_or(0));
             }
             black_box(sum)
         })
@@ -18,9 +18,9 @@ fn bench_reader(c: &mut Criterion) {
     c.bench_function("reader_read_bits_8", |b| {
         b.iter(|| {
             let mut reader = Reader::new(black_box(&data), 0, data.len());
-            let mut sum = 0;
+            let mut sum = 0u64;
             for _ in 0..1000 {
-                sum += reader.read_bits(8).unwrap_or(0);
+                sum += u64::from(reader.read_bits(8).unwrap_or(0));
             }
             black_box(sum)
         })
@@ -29,9 +29,9 @@ fn bench_reader(c: &mut Criterion) {
     c.bench_function("reader_read_bits_32", |b| {
         b.iter(|| {
             let mut reader = Reader::new(black_box(&data), 0, data.len());
-            let mut sum = 0;
+            let mut sum = 0u64;
             for _ in 0..1000 {
-                sum += reader.read_bits(32).unwrap_or(0);
+                sum += u64::from(reader.read_bits(32).unwrap_or(0));
             }
             black_box(sum)
         })
@@ -40,9 +40,9 @@ fn bench_reader(c: &mut Criterion) {
     c.bench_function("reader_read_bits_13", |b| {
         b.iter(|| {
             let mut reader = Reader::new(black_box(&data), 0, data.len());
-            let mut sum = 0;
+            let mut sum = 0u64;
             for _ in 0..1000 {
-                sum += reader.read_bits(13).unwrap_or(0);
+                sum += u64::from(reader.read_bits(13).unwrap_or(0));
             }
             black_box(sum)
         })
